MNT: move toy classes to examples
These are toy/testing classes that yield synthetic data, they should
not be exposed as part of the core run_engine file.
